pyproject.toml
[tool.poetry]
name = "echo-client"
version = "0.3.0"
description = ""
authors = ["Rickyxrc <rickyxrc@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4"
websockets = "^12.0"
rich = "^13.7.0"
prompt-toolkit = "^3.0.43"
pypinyin = "^0.50.0"
pyyaml = "^6.0.1"
jieba = "^0.42.1"
"markdown-it-py" = "^3.0.0"

[tool.poetry.scripts]
echo-client = 'main:main'

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.3"
pre-commit = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
